api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai[realtime]>=1.54.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
